backend/requirements.txt
fastapi
groq
uvicorn
python-dotenv
google-generativeai
langgraph
langchain
langchain-google-genai
slack-sdk
google-auth
google-auth-oauthlib
google-api-python-client
reportlab
pyaudio
speechrecognition
python-multipart
websockets
httpx
